+++ test/unit/promptEdit.unit.test.js
@@ -1,4 +1,4 @@
-import { makePromptEditable } from "../script/promptEdit.js";
+import { makePromptEditable } from "../../script/promptEdit.js";
 import { test, expect } from "@jest/globals";
 
 test("makePromptEditable doesn't return a value", () => {
